+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/AuthTests.java
@@ -90,6 +90,7 @@ public class AuthTests {
             Thread.sleep(2000);
         } catch (InterruptedException e) {
             e.printStackTrace();
+            fail();
         }
         looperThread.testComplete();
     }