+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * All major public classes are now non-final. This is mostly a compromise to support Mockito. All protected fields/methods are still not considered part of the public API and can change without notice (#3869).
 * All Realm instances share a single notification daemon thread.
 * Fixed Java lint warnings with generated proxy classes (#2929).
+* Add 'like' predicate for String fields (#3752)
 
 ### Internal