+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ### Internal
 
-* Updated Realm Sync to 1.0.0-BETA-7.0.
+* Updated Realm Sync to 1.0.0-BETA-7.1.
 
 ## 2.2.2